Fix import of OptionsManager for firedrake release (#198)
Fixes #196

The fix to import OptionsManager from petsctools, which is needed in
firedrake master, appears to break with the current firedrake release
version because petsctools needs to be initialized (which is done by
firedrake) - also release doesn't depend on petsctools so the package
isn't installed automatically.

Attempt to import OptionsManager directly from firedrake first - which
should work in release - and if not import it from petsctools.